+++ include/gque.h
@@ -11,8 +11,10 @@
 
 namespace gcl {
 
-//--- generic cuicular queue
-class gcl_api circular_queue {
+//--------------------------------------------------------------------
+//	class CircularQueue
+//--------------------------------------------------------------------
+class gcl_api CircularQueue {
 protected:
     typedef struct {
         unsigned begin;     // constant: start of que (end of que contro block)
@@ -26,9 +28,9 @@ protected:
     // template<class T> friend class que;
 
 public:
-    circular_queue() {}
-    circular_queue(unsigned capacity, unsigned item_size) { init(capacity, item_size); }
-    ~circular_queue();
+    CircularQueue() {}
+    CircularQueue(unsigned capacity, unsigned item_size) { init(capacity, item_size); }
+    ~CircularQueue();
 
 	bool init(unsigned capacity, unsigned item_size);
     void clear() const { if (m_q) { m_q->head = m_q->tail = m_q->begin; } }
@@ -64,19 +66,21 @@ protected:
 };
 
 
-//--- typed (circular) queue
+//--------------------------------------------------------------------
+//	class Queue - typed (circular) queue
+//--------------------------------------------------------------------
 template <class T>
-class gcl_api queue : public circular_queue {
+class gcl_api Queue : public CircularQueue {
 public:
-    queue() {}
-    queue(unsigned capacity): circular_queue(capacity, sizeof(T)) {}
+    Queue() {}
+    Queue(unsigned capacity): CircularQueue(capacity, sizeof(T)) {}
 
-    bool init(unsigned capacity) { return circular_queue::init(capacity, sizeof(T)); };
+    bool init(unsigned capacity) { return CircularQueue::init(capacity, sizeof(T)); };
 
-	bool put(const T *item=0) const { return circular_queue::put(item); }
-	bool get(T *item=0) const { return circular_queue::get(item); }
-	bool push(const T *item=0) const { return circular_queue::push(item); }
-	bool pop(T *item=0) const { return circular_queue::pop(item); }
+	bool put(const T *item=0) const { return CircularQueue::put(item); }
+	bool get(T *item=0) const { return CircularQueue::get(item); }
+	bool push(const T *item=0) const { return CircularQueue::push(item); }
+	bool pop(T *item=0) const { return CircularQueue::pop(item); }
 
     //--- support reference type interface
     bool put(const T &item) const { return put(&item); }
@@ -84,36 +88,39 @@ public:
 	bool push(const T &item) const { return push(&item); }
 	bool pop(T &item) const { return pop(&item); }
 
-    T *first() const { return (T *)circular_queue::first(); }
-    T *last() const { return (T *)circular_queue::last(); }
-    T *peek() const { return (T *)circular_queue::peek(); }
-	T *peek_next(const T *peek=0) const { return (T *)circular_queue::peek_next(peek); }
-	T *peek_prev(const T *peek=0) const { return (T *)circular_queue::peek_prev(peek); }
+    T *first() const { return (T *)CircularQueue::first(); }
+    T *last() const { return (T *)CircularQueue::last(); }
+    T *peek() const { return (T *)CircularQueue::peek(); }
+	T *peek_next(const T *peek=0) const { return (T *)CircularQueue::peek_next(peek); }
+	T *peek_prev(const T *peek=0) const { return (T *)CircularQueue::peek_prev(peek); }
 
 	//--- accessors
-	T *head() const { return (T *)circular_queue::head(); }
-	T *tail() const { return (T *)circular_queue::tail(); }
-	T *begin() const { return (T *)circular_queue::begin(); }
-	T *end() const { return (T *)circular_queue::end(); }
+	T *head() const { return (T *)CircularQueue::head(); }
+	T *tail() const { return (T *)CircularQueue::tail(); }
+	T *begin() const { return (T *)CircularQueue::begin(); }
+	T *end() const { return (T *)CircularQueue::end(); }
 };
 
 
-//--- thread-safe version of queue (mt: multi-threading)
+
+//--------------------------------------------------------------------
+//	class Queue - thread-safe version of Queue (MT: multi-threading)
+//--------------------------------------------------------------------
 template <class T>
-class gcl_api queue_mt : protected queue<T> {
-    rwlock       m_lock;
+class gcl_api QueueMT : protected Queue<T> {
+    RWLock       m_lock;
 
 public:
-    queue_mt() {}
-    queue_mt(unsigned capacity): queue<T>(capacity) {}
+    QueueMT() {}
+    QueueMT(unsigned capacity): Queue<T>(capacity) {}
 
-    bool init(unsigned capacity) { autolock_wr lock(m_lock); return queue<T>::init(capacity, sizeof(T)); };
-    void clear() const { autolock_wr lock(m_lock); queue<T>::clear(); }
+    bool init(unsigned capacity) { AutoLockWrite lock(m_lock); return Queue<T>::init(capacity, sizeof(T)); };
+    void clear() const { AutoLockWrite lock(m_lock); Queue<T>::clear(); }
 
-    bool put(const T *item=0) { autolock_wr lock(m_lock); return queue<T>::put(item); }
-    bool get(T *item=0) { autolock_wr lock(m_lock); return queue<T>::get(item); }
-    bool push(const T *item=0) { autolock_wr lock(m_lock); return queue<T>::push(item); }
-	bool pop(T *item=0) { autolock_wr lock(m_lock); return queue<T>::get(item); };
+    bool put(const T *item=0) { AutoLockWrite lock(m_lock); return Queue<T>::put(item); }
+    bool get(T *item=0) { AutoLockWrite lock(m_lock); return Queue<T>::get(item); }
+    bool push(const T *item=0) { AutoLockWrite lock(m_lock); return Queue<T>::push(item); }
+	bool pop(T *item=0) { AutoLockWrite lock(m_lock); return Queue<T>::get(item); };
 
     //--- support reference type interface
     bool put(const T &item) { return put(&item); }
@@ -122,13 +129,13 @@ public:
 	bool pop(T &item) { return pop(&item); }
 
 	//--- accessors
-	unsigned item_size() { autolock_rd lock(m_lock); return queue<T>::item_size(); }
+	unsigned item_size() { AutoLockRead lock(m_lock); return Queue<T>::item_size(); }
 
-    bool is_empty() { autolock_rd lock(m_lock); return queue<T>::is_empty(); }
-    bool is_full() { autolock_rd lock(m_lock); return queue<T>::is_full(); }
-	unsigned length() { autolock_rd lock(m_lock); return queue<T>::length(); }
-    unsigned available() { autolock_rd lock(m_lock); return queue<T>::available(); }
-    unsigned capacity() { autolock_rd lock(m_lock); return queue<T>::capacity(); }
+    bool is_empty() { AutoLockRead lock(m_lock); return Queue<T>::is_empty(); }
+    bool is_full() { AutoLockRead lock(m_lock); return Queue<T>::is_full(); }
+	unsigned length() { AutoLockRead lock(m_lock); return Queue<T>::length(); }
+    unsigned available() { AutoLockRead lock(m_lock); return Queue<T>::available(); }
+    unsigned capacity() { AutoLockRead lock(m_lock); return Queue<T>::capacity(); }
 };
 
 } // namespace gcl