+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - Update pthread WIN32 port to use pthreads4w/3.0.0 (downloaded using conan)
 - Update naming convention referencing python style guideline
 - Improve high precision time functions
-- Redesign queue, list, event_emitter
-- Add autolock classes
+- Redesign Queue, List, EventEmitter
+- Add AutoLock classes
 
 ## Tested platforms
 - gcc 8.3 (x86 and x64)